# apps/daemon/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageDocModel": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "frontmatter": {
            "additionalProperties": true,
            "title": "Frontmatter",
            "type": "object"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "path",
          "id",
          "title",
          "icon",
          "frontmatter",
          "body",
          "hash"
        ],
        "title": "PageDocModel",
        "type": "object"
      },
...
      "PreviewBody": {
        "properties": {
          "cfg": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cfg"
          },
          "exaggeration": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exaggeration"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "text": {
            "default": "",
            "maxLength": 300,
            "title": "Text",
            "type": "string"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          }
        },
        "title": "PreviewBody",
        "type": "object"
      },
...
      "PurgeResultModel": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          }
        },
        "required": [
          "entries",
          "bytes"
        ],
        "title": "PurgeResultModel",
        "type": "object"
      },
...
      "RenameVoice": {
        "properties": {
          "name": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenameVoice",
        "type": "object"
      },
...
      "RestoreRequest": {
        "properties": {
          "target_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Path"
          }
        },
        "title": "RestoreRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoiceInfo": {
        "properties": {
          "consent_at": {
            "title": "Consent At",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "seconds": {
            "title": "Seconds",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "seconds",
          "created_at",
          "consent_at"
        ],
        "title": "VoiceInfo",
        "type": "object"
      },
      "VoiceList": {
        "properties": {
          "voices": {
            "items": {
              "$ref": "#/components/schemas/VoiceInfo"
            },
            "title": "Voices",
            "type": "array"
          }
        },
        "required": [
          "voices"
        ],
        "title": "VoiceList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "graite-daemon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trash/{trash_id}": {
      "delete": {
        "description": "Permanently delete one trash entry. This cannot be undone.",
        "operationId": "purge_trash_entry_api_v1_trash__trash_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "trash_id",
            "required": true,
            "schema": {
              "title": "Trash Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge Trash Entry",
        "tags": [
          "trash"
        ]
      }
    },
    "/api/v1/trash/{trash_id}/restore": {
      "post": {
        "operationId": "restore_trash_api_v1_trash__trash_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "trash_id",
            "required": true,
            "schema": {
              "title": "Trash Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RestoreRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDocModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Trash",
        "tags": [
          "trash"
        ]
      }
    },
...
    "/api/v1/voice/preview": {
      "post": {
        "description": "One sentence in a voice, as a WAV. The engine stays loaded for a minute afterwards.",
        "operationId": "preview_api_v1_voice_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "Voice"
        ]
      }
    },
...
    "/api/v1/voice/voices": {
      "get": {
        "operationId": "list_voices_api_v1_voice_voices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Voices",
        "tags": [
          "Voice"
        ]
      },
      "post": {
        "description": "A voice from a short recording of one person (the request body). It is converted to\nthe mono 24 kHz WAV the voice engine reads, and only stored with the confirmation that\nthe user may use this voice.",
        "operationId": "add_voice_api_v1_voice_voices_post",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "consent",
            "required": false,
            "schema": {
              "default": false,
              "title": "Consent",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "default": "voice.wav",
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Voice",
        "tags": [
          "Voice"
        ]
      }
    },
    "/api/v1/voice/voices/{voice_id}": {
      "delete": {
        "operationId": "delete_voice_api_v1_voice_voices__voice_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "voice_id",
            "required": true,
            "schema": {
              "title": "Voice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Voice Api V1 Voice Voices  Voice Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Voice",
        "tags": [
          "Voice"
        ]
      },
      "patch": {
        "operationId": "rename_voice_api_v1_voice_voices__voice_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "voice_id",
            "required": true,
            "schema": {
              "title": "Voice Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameVoice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Voice",
        "tags": [
          "Voice"
        ]
      }
    },
...
  }
}